requirements.txt
FastAPI[standard]
python-dotenv
mysql-connector-python
PyJWT
bcrypt